[simple-memo] refactor : 메서드 이름 수정, 도메인 모델 패키지 분리
