+++ simplememo/simplememo/src/main/java/gunhee/simplememo/domain/memo/IncomeExpenseType.java
@@ -1,4 +1,4 @@
-package gunhee.simplememo.domain;
+package gunhee.simplememo.domain.memo;
 
 public enum IncomeExpenseType {
     INCOME, EXPENSE
