+++ simplememo/simplememo/src/main/java/gunhee/simplememo/dto/memo/MemoResponse.java
@@ -1,8 +1,8 @@
 package gunhee.simplememo.dto.memo;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import gunhee.simplememo.domain.IncomeExpenseType;
-import gunhee.simplememo.domain.Memo;
+import gunhee.simplememo.domain.memo.IncomeExpenseType;
+import gunhee.simplememo.domain.memo.Memo;
 import lombok.Getter;
 
 import java.math.BigDecimal;
